requirements.txt
requests>=2.28.0
pyyaml>=5.0
